requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
openai==1.12.0
python-dotenv==1.0.0
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0
nltk==3.8.1
selenium==4.16.0
webdriver-manager==4.0.1
Pillow==10.2.0
gunicorn==21.2.0
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.2.2
chromadb==0.4.22 